remove obsolete files and move other into ack folder
