+++ src/ack/code.ts
@@ -1,4 +1,4 @@
-export enum ActionResultCode {
+export enum AckCode {
 	NOT_SET = 0,
 	SUCCESS = 1,
 	FAILURE = -1
