+++ src/ack/options.ts
@@ -1,4 +1,4 @@
-export interface ActionResultOptions<T = unknown> {
+export interface AckOptions<T = unknown> {
 	errorThreshold?: number;
 	payload?: T;
 	serialized?: string;
