+++ src/index.ts
@@ -1,4 +1,4 @@
-export * from './action-result';
-export * from './action-result/code';
-export * from './action-result/options';
-export * from './action-result/state';
+export * from './ack';
+export * from './ack/code';
+export * from './ack/options';
+export * from './ack/state';
